Refactored out parsing from level creation I refactored the code into a super class because I want to generate levels that don't have to be based on files in a directory.
