+++ src/test/java/level/LevelCreationStepDefHelper.java
@@ -1,4 +1,4 @@
-package parser;
+package level;
 
 import java.io.FileNotFoundException;
 import java.io.PrintWriter;
