+++ src/test/java/level/LevelCreatorITHelper.java
@@ -1,4 +1,4 @@
-package parser;
+package level;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
@@ -49,7 +49,7 @@ public class LevelCreatorITHelper {
 	}
 
 	protected void createLevel() throws Throwable {
-		LevelCreator levelCreator = new LevelCreator(TestingTunableParameters.FILE_LOCATION_PREFIX,
+		LevelCreator levelCreator = new FileParserLevelCreator(TestingTunableParameters.FILE_LOCATION_PREFIX,
 				new ReaderWrapper());
 		try {
 			gameEngine = new GameEngine(levelCreator);
@@ -91,7 +91,7 @@ public class LevelCreatorITHelper {
 		BufferedReader bufferedReader = Mockito.mock(BufferedReader.class);
 		Mockito.when(readerWrapper.createBufferedReader(Mockito.anyString())).thenReturn(bufferedReader);
 		Mockito.doThrow(ioException).when(bufferedReader).readLine();
-		LevelCreator levelCreator = new LevelCreator(TestingTunableParameters.FILE_LOCATION_PREFIX, readerWrapper);
+		LevelCreator levelCreator = new FileParserLevelCreator(TestingTunableParameters.FILE_LOCATION_PREFIX, readerWrapper);
 		gameEngine = new GameEngine(levelCreator);
 	}
