+++ src/test/java/level/LevelCreatorIntegrationTest.java
@@ -1,4 +1,4 @@
-package parser;
+package level;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
