+++ src/test/java/level/LevelCreatorTest.java
@@ -1,4 +1,4 @@
-package parser;
+package level;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
@@ -18,7 +18,7 @@ import wrappers.ReaderWrapper;
 public class LevelCreatorTest {
 
 	private GameEngine gameEngine;
-	private LevelCreator levelCreator;
+	private FileParserLevelCreator levelCreator;
 	private final String FILE_LOCATION_PREFIX = "FILE_LOCATION_PREFIX";
 	private ReaderWrapper readerWrapper;
 	private final int LEVEL = 1;
@@ -27,7 +27,7 @@ public class LevelCreatorTest {
 	public void setUp() {
 		gameEngine = Mockito.mock(GameEngine.class);
 		readerWrapper = Mockito.mock(ReaderWrapper.class);
-		levelCreator = new LevelCreator(FILE_LOCATION_PREFIX, readerWrapper);
+		levelCreator = new FileParserLevelCreator(FILE_LOCATION_PREFIX, readerWrapper);
 	}
 
 	@Test
