+++ src/test/java/main/MovementStepDefs.java
@@ -9,8 +9,9 @@ import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
 import engine.GameEngine;
-import parser.LevelCreationStepDefHelper;
-import parser.LevelCreator;
+import level.FileParserLevelCreator;
+import level.LevelCreationStepDefHelper;
+import level.LevelCreator;
 import values.TestingTunableParameters;
 import wrappers.ReaderWrapper;
 
@@ -22,7 +23,7 @@ public class MovementStepDefs extends LevelCreationStepDefHelper {
 	public void level_is(List<String> levelStrings) throws Throwable {
 		writeLevelFile(levelStrings);
 		gameEngine = new GameEngine(
-				new LevelCreator(TestingTunableParameters.FILE_LOCATION_PREFIX, new ReaderWrapper()));
+				new FileParserLevelCreator(TestingTunableParameters.FILE_LOCATION_PREFIX, new ReaderWrapper()));
 	}
 
 	@When("^the player moves left$")
